requirements.txt
click==7.1.2
fastapi==0.63.0
Flask==1.1.2
gunicorn==20.1.0
h11==0.12.0
handler==1.3.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pydantic==1.8.1
starlette==0.13.6
typing-extensions==3.7.4.3
unicorn==1.0.2
urlparser==0.1.2
uvicorn==0.13.4
Werkzeug==1.0.1
